refactor(typebot-mcp-py): split into transport / services / tools layers
Replace the 759-LOC server.py monolith and stateful TypebotClient with a
three-layer architecture grounded in FastMCP idioms:

- transport.py — pure async HTTP helper (build_headers + request)
- services/ — stateless domain wrappers, one module per upstream area
  (chat, typebots, results, analytics, folders, workspaces)
- tools/ — MCP tool definitions, one module per domain, registered via
  register_all(mcp, cfg, app); writes gated by allow_writes
- context.py + lifespan.py — AppContext dataclass with both httpx
  clients, opened eagerly and closed by FastMCP lifespan
- errors.py — http_errors_as_tool_errors adapter so TypebotHTTPError
  becomes mcp.ToolError (protocol-correct isError: true) instead of
  the previous {"ok": False, ...} envelope

Every tool now declares ToolAnnotations(readOnlyHint, destructiveHint,
idempotentHint); the table is enforced by tests/test_annotations.py.

Tests reorganised: client tests dropped, replaced by service-level
respx tests (test_services_chat, test_services_management), plus new
test_lifespan, test_errors, and test_annotations. test_server now
asserts ToolError is raised on HTTP failure instead of inspecting an
envelope payload.

CLAUDE.md and README.md rewritten to match. Both gain an explicit
instruction to update them in the same change to prevent drift.

Verified end-to-end against a live Typebot instance (list_workspaces,
list_typebots, get_typebot, get_published_typebot, start_chat,
continue_chat, start_chat_preview, get_analytics_stats, list_results,
list_folders, list_credentials all return expected payloads).

Pipeline green: ruff check, ruff format, pyright (0 errors), pytest
(42 passed).
