+++ packages/typebot-mcp-py/src/typebot_mcp/__init__.py
@@ -1,6 +1,5 @@
 """Typebot MCP — Python Model Context Protocol server for Typebot."""
 
-from typebot_mcp.client import TypebotClient
 from typebot_mcp.config import Settings
 from typebot_mcp.exceptions import TypebotError, TypebotHTTPError
 from typebot_mcp.server import build_server
@@ -9,7 +8,6 @@ __version__ = "0.1.0"
 
 __all__ = [
     "Settings",
-    "TypebotClient",
     "TypebotError",
     "TypebotHTTPError",
     "__version__",
