+++ packages/typebot-mcp-py/src/typebot_mcp/services/__init__.py
@@ -1,0 +1,1 @@
+"""Stateless domain-grouped wrappers around Typebot's REST endpoints."""
